+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - add language check `condition`
  - add an option for multiple selection of jobs in `scheduledJobs` and then `execute` them or `delete` them
  - for `formActions` list on right click open edit or view action in a new tab
+ - add `searchBuilder` for `formActions`;
 
 # v5.1.5
 ### Bugfix